Now automatically connects to the Wallabies.
You no longer need to connect to the Wallabies manually, the program creates a new profile for the Wallabies based on the serial numbers and connects itself. This is done by calculating the sha256 hash of the serial and then taking the first 6 digits of the output and adding 00 to the end (the way the wallaby password is generated).